% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppHandlers.R
\name{colKurt}
\alias{colKurt}
\title{Computes sample excess kurtosis for each column of a matrix}
\usage{
colKurt(x)
}
\arguments{
\item{x}{Numeric Matrix}
}
\value{
Vector of sample excess kurtosis for all columns
}
\description{
Computes sample excess kurtosis for each column of a matrix
}
