% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colMax}
\alias{colMax}
\title{Computes columnwise maximum of a numeric matrix}
\usage{
colMax(x)
}
\arguments{
\item{x}{Numeric Matrix}
}
\value{
Vector of maximum value of each column
}
\description{
Computes columnwise maximum of a numeric matrix
}
