% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppHandlers.R
\name{colMin}
\alias{colMin}
\title{Computes columnwise minimum of a numeric matrix}
\usage{
colMin(x)
}
\arguments{
\item{x}{Numeric Matrix}
}
\value{
Vector of minimum value of each column
}
\description{
Computes columnwise minimum of a numeric matrix
}
