% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colQuant}
\alias{colQuant}
\title{Computes columnwise quantiles of a numeric matrix}
\usage{
colQuant(a, probs)
}
\arguments{
\item{probs}{Vector of quantiles to compute}

\item{x}{Numeric Matrix}
}
\value{
Vector/Matrix of columnwise quantiles
}
\description{
Computes columnwise quantiles of a numeric matrix
}
