% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppHandlers.R
\name{colSkew}
\alias{colSkew}
\title{Computes sample skewness for each column of a matrix}
\usage{
colSkew(x)
}
\arguments{
\item{x}{Numeric Matrix}
}
\value{
Vector of sample skewness for all columns
}
\description{
Computes sample skewness for each column of a matrix
}
