% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{colVar}
\alias{colVar}
\title{Computes rowwise sample variances}
\usage{
colVar(x)
}
\arguments{
\item{x}{Numeric Matrix}
}
\value{
Vector of sample variances of each row
}
\description{
Computes rowwise sample variances
}
