% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linReg.R
\name{lmBoot}
\alias{lmBoot}
\title{Compute bootstrapped linear regression}
\usage{
lmBoot(y, X, nboot = 100, intercept = T, boot.type = "np")
}
\arguments{
\item{y}{Numeric Vector of dependent variables}

\item{X}{Numeric Matrix or data.frame of dim n x k, where n is the number of observations and k the number of independent variables}

\item{nboot}{Number of bootstrap iterations (default is 100)}

\item{intercept}{Logical, specifying whether an intercept should be used (default is TRUE)}

\item{boot.type}{Character string specifying which bootstrap type should be used. Must be either "np" (Non-parametric, default) or "wild" (Wild Rademacher bootstrap)}
}
\value{
Vector of sample skewness for all rows
}
\description{
Compute bootstrapped linear regression
}
