% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_whitney_test.R
\name{mann.whitney.testBoot}
\alias{mann.whitney.testBoot}
\title{Resampling version of the student t-Test for one or two samples}
\usage{
mann.whitney.testBoot(
  x,
  y,
  alpha = 0.05,
  alternative = "two.sided",
  nboot = 100,
  boot.type = "perm"
)
}
\arguments{
\item{x, y}{Sample Data (Non-empty numeric vectors)}

\item{alpha}{Significance level (default is 0.05).}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{nboot}{Integer specifying the number of bootstrap iterations (default is 100)}

\item{boot.type}{A character string specifying the bootstrap type, must be one of "perm" (Permutation, default) or "np" (Non-parametric bootstrap).}
}
\value{
A list of class "htest"
}
\description{
Resampling version of the student t-Test for one or two samples
}
