% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppHandlers.R
\name{matKurtosis}
\alias{matKurtosis}
\title{Computes row- or columnwise sample kurtosis.}
\usage{
matKurtosis(x, margin = 1)
}
\arguments{
\item{x}{Numeric Matrix or data.frame.}

\item{margin}{Scalar giving the subscripts which the function will be applied over, such as in apply. 1 indicates rows (default), 2 indicates columns.}
}
\value{
Vector of sample kurtosis of each row/column.
}
\description{
Computes row- or columnwise sample kurtosis.
}
