% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppHandlers.R
\name{matMinMax}
\alias{matMinMax}
\title{Computes row- or columnwise Minimum or Maximum.}
\usage{
matMinMax(x, margin = 1, min = T)
}
\arguments{
\item{x}{Numeric Matrix or data.frame.}

\item{margin}{Scalar giving the subscripts which the function will be applied over, such as in apply. 1 indicates rows (default), 2 indicates columns.}

\item{min}{Logical, whether the Minimum should be computed (default is TRUE), otherwise Maximum is computed (FALSE).}
}
\value{
Vector of Min/Max of each row/column.
}
\description{
Computes row- or columnwise Minimum or Maximum.
}
