% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppHandlers.R
\name{matQuantile}
\alias{matQuantile}
\title{Computes row- or columnwise quantiles. The computation mimics the default computation of stat's quantile function (type 7).}
\usage{
matQuantile(x, margin = 1, probs = 0.5)
}
\arguments{
\item{x}{Numeric Matrix or data.frame.}

\item{margin}{Scalar giving the subscripts which the function will be applied over, such as in apply. 1 indicates rows (default), 2 indicates columns.}

\item{probs}{Vector of quantiles to compute, default is 0.5.}
}
\value{
Vector of row-/columnwise quantiles specified in probs.
}
\description{
Computes row- or columnwise quantiles. The computation mimics the default computation of stat's quantile function (type 7).
}
