% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppHandlers.R
\name{matRank}
\alias{matRank}
\title{Computes row- or columnwise ranks. Ties are adjusted with mean-values of given ranks (statistical tie handling).}
\usage{
matRank(x, margin = 1)
}
\arguments{
\item{x}{Numeric Matrix or data.frame.}

\item{margin}{Scalar giving the subscripts which the function will be applied over, such as in apply. 1 indicates rows (default), 2 indicates columns.}
}
\value{
Vector of row-/columnwise ranks.
}
\description{
Computes row- or columnwise ranks. Ties are adjusted with mean-values of given ranks (statistical tie handling).
}
