% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{matVar}
\alias{matVar}
\title{Computes row- or columnwise sample variances.}
\usage{
matVar(x, margin = 1)
}
\arguments{
\item{x}{Numeric Matrix or data.frame.}

\item{margin}{Scalar giving the subscripts which the function will be applied over, such as in apply. 1 indicates rows, 2 indicates columns.}
}
\value{
Vector of sample variances of each row/column.
}
\description{
Computes row- or columnwise sample variances.
}
