% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowKurt}
\alias{rowKurt}
\title{Computes sample excess kurtosis for each row of a matrix}
\usage{
rowKurt(x)
}
\arguments{
\item{x}{Numeric Matrix}
}
\value{
Vector of sample excess kurtosis for all rows
}
\description{
Computes sample excess kurtosis for each row of a matrix
}
