% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppHandlers.R
\name{rowMax}
\alias{rowMax}
\title{Computes rowwise maximum of a numeric matrix}
\usage{
rowMax(x)
}
\arguments{
\item{x}{Numeric Matrix}
}
\value{
Vector of maximum value of each row
}
\description{
Computes rowwise maximum of a numeric matrix
}
