% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowMin}
\alias{rowMin}
\title{Computes rowwise minimum of a numeric matrix}
\usage{
rowMin(x)
}
\arguments{
\item{x}{Numeric Matrix}
}
\value{
Vector of minimum value of each row
}
\description{
Computes rowwise minimum of a numeric matrix
}
