% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowQuant}
\alias{rowQuant}
\title{Computes rowwise quantiles of a numeric matrix}
\usage{
rowQuant(a, probs)
}
\arguments{
\item{probs}{Vector of quantiles to compute}

\item{x}{Numeric Matrix}
}
\value{
Vector/Matrix of rowwise quantiles
}
\description{
Computes rowwise quantiles of a numeric matrix
}
