% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppHandlers.R
\name{rowSkew}
\alias{rowSkew}
\title{Computes sample skewness for each row of a matrix}
\usage{
rowSkew(x)
}
\arguments{
\item{x}{Numeric Matrix}
}
\value{
Vector of sample skewness for all rows
}
\description{
Computes sample skewness for each row of a matrix
}
