% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t.testBoot}
\alias{t.testBoot}
\title{Bootstrap version of the student t-Test for one or two samples}
\usage{
\method{t}{testBoot}(x, y = NULL, mu.0 = 0, alpha = 0.05,
  alternative = "two.sided", nboot = 100, boot.type = "np")
}
\arguments{
\item{x}{Sample Data (Non-empty numeric vector)}

\item{y}{Optional Non-empty numeric vector to compare against (two-sample-test).}

\item{mu.0}{Expected value under the null hypothesis (one-sample-test), default is 0.}

\item{alpha}{Significance level (default is 0.05).}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{nboot}{Integer specifying the number of bootstrap iterations (default is 100)}

\item{boot.type}{A character string specifying the bootstrap type, must be one of "np" (Non-parametric bootstrap, default), "wild" (Rademacher wild bootstrap) or "npg" (groupwise non-parametric, two-sample only).}
}
\value{
A list with class "htest" containing the following components
}
\description{
Bootstrap version of the student t-Test for one or two samples
}
