% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppHandlers.R
\name{colQuant}
\alias{colQuant}
\title{Computes columnwise quantiles of a numeric matrix}
\usage{
colQuant(x, probs = 0.5)
}
\arguments{
\item{x}{Numeric Matrix}

\item{probs}{Vector of quantiles to compute}
}
\value{
Vector/Matrix of columnwise quantiles
}
\description{
Computes columnwise quantiles of a numeric matrix
}
