% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppHandlers.R
\name{rowQuant}
\alias{rowQuant}
\title{Computes rowwise quantiles of a numeric matrix}
\usage{
rowQuant(x, probs = 0.5)
}
\arguments{
\item{x}{Numeric Matrix}

\item{probs}{Vector of quantiles to compute}
}
\value{
Vector/Matrix of rowwise quantiles
}
\description{
Computes rowwise quantiles of a numeric matrix
}
